requirements.txt
django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-rest-swagger==2.2.0
drf-yasg==1.21.7
ipdb==0.13.13
pyyaml==6.0.1
psycopg2-binary==2.9.9